+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## Version 0.0.6
+- add LngLat#approxEquals 
+
+## Version 0.0.5
+- modify polyline#project return value
+
+## Version 0.0.4
+- add Earth#getInitalBearingDegree method for initialBearing
+
+## Verison 0.0.3
+- add Earth support
+- add polyline#getCentroid and polyline#interpolate and polyline#project method
+
 ## Verison 0.0.2
 
 - Introduce s2.RegionCoverer.getRadiusCovering methods that take a lat/lng and radius and generates a covering